refactor(pod): remove pod logs handler and integrate into aggregator handler
