+++ pkg/core/handler/aggregator/log.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package pod
+package aggregator
 
 import (
 	"bufio"
