+++ ui/src/pages/insightDetail/components/podLogs/index.tsx
@@ -70,7 +70,7 @@ const PodLogs: React.FC<PodLogsProps> = ({
       setLogs([]) // Clear logs when switching containers or reconnecting
     }
 
-    const url = `/rest-api/v1/pods/${cluster}/${namespace}/${podName}/logs?container=${container}`
+    const url = `/rest-api/v1/insight/aggregator/pod/${cluster}/${namespace}/${podName}/log?container=${container}`
     const eventSource = new EventSource(url)
     eventSourceRef.current = eventSource
